+++ SW1_ISO9126_FUZZY/Vistas/FormularioEvaluacionPage.xaml.cs
@@ -1,273 +1,349 @@
-﻿using Newtonsoft.Json;
-using SW1_ISO9126_FUZZY.JSON;
+﻿using SW1_ISO9126_FUZZY.JSON;
 using SW1_ISO9126_FUZZY.Modelo_Datos;
-using SW1_ISO9126_FUZZY.Modelo_Datos.Etiquetas;
-using System;
+using SW1_ISO9126_FUZZY.Modelo_Datos.Listas;
 using System.Collections;
-using System.IO;
 using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Media;
+using System.Windows.Navigation;
 
 namespace SW1_ISO9126_FUZZY.Vistas
 {
-    /// <summary>
-    /// Lógica de interacción para FormularioEvaluacionPage.xaml
-    /// </summary>
-    public partial class FormularioEvaluacionPage : Page {
+	/// <summary>
+	/// Lógica de interacción para FormEvaluacionPage.xaml
+	/// </summary>
+	public partial class FormEvaluacionPage : Page
+	{
+        // Listas metricas y seleccion metrica
+        private ArrayList listaMetricas;
+        private ArrayList listaSeleccion;
+        private ArrayList listaEvaluacion;
+        private int indiceListas;
 
-        // Caracteristicas
-        private JFuncionabilidad funInt;
-        private JFuncionabilidad funExt;
-        private JUsabilidad usaInt;
-        private JUsabilidad usaExt;
-        private JMantenibilidad manInt;
-        private JMantenibilidad manExt;
+        private FormularioEvaluacionPage origen;
 
-        // Estado caracteristicas
-        private bool isFunIntAct;
-        private bool isFunExtAct;
-        private bool isUsaIntAct;
-        private bool isUsaExtAct;
-        private bool isManIntAct;
-        private bool isManExtAct;
-
-        // Listas de caracteristicas
-        private ArrayList funcionalidadInterna;
-        private ArrayList funcionalidadExterna;
-        private ArrayList usabilidadInterna;
-        private ArrayList usabilidadExterna;
-        private ArrayList mantenibilidadInterna;
-        private ArrayList mantenibilidadExterna;
-
-        // Listas de caracteristicas seleccionadas
-        private ArrayList MTSfuncionalidadInterna;
-        private ArrayList MTSfuncionalidadExterna;
-        private ArrayList MTSusabilidadInterna;
-        private ArrayList MTSusabilidadExterna;
-        private ArrayList MTSmantenibilidadInterna;
-        private ArrayList MTSmantenibilidadExterna;
-
-        // Listas de caracteristicas evaluadas
-        private ArrayList MTEfuncionalidadInterna;
-        private ArrayList MTEfuncionalidadExterna;
-        private ArrayList MTEusabilidadInterna;
-        private ArrayList MTEusabilidadExterna;
-        private ArrayList MTEmantenibilidadInterna;
-        private ArrayList MTEmantenibilidadExterna;
-
-        private FormEvaluacionPage paginaEvaluacion;
-        /*private Respuesta metricas;
-        private EstadoModulo evalMetricas;*/
-        private Evaluacion miEvaluacion;
-
-        public FormularioEvaluacionPage(Evaluacion nueva) {
-
+        public FormEvaluacionPage()
+        {
             InitializeComponent();
-            this.paginaEvaluacion = new FormEvaluacionPage();
-           /* this.metricas = new Respuesta();
-            this.evalMetricas = new EstadoModulo();*/
-            this.miEvaluacion = nueva;
+            cargarEntorno();
+		}
 
-            inicializarEstadoCaracteristica();
+        // Metodos
+
+        private void cargarEntorno()
+        {
+            inicializarBotones();
             inicializarListas();
-            inicializarSeleccion();
-            inicializarEvaluacion();
-            cargarJsonMetricas();
         }
 
-        private void inicializarEstadoCaracteristica()
+        private void inicializarBotones()
         {
-            this.isFunIntAct = false;
-            this.isFunExtAct = false;
-            this.isUsaIntAct = false;
-            this.isUsaExtAct = false;
-            this.isManIntAct = false;
-            this.isManExtAct = false;
+            btnAnterior.IsEnabled = false;
+            btnSiguiente.IsEnabled = true;
         }
 
         private void inicializarListas()
         {
-            this.funcionalidadInterna = new ArrayList();
-            this.funcionalidadExterna = new ArrayList();
-            this.usabilidadInterna = new ArrayList();
-            this.usabilidadExterna = new ArrayList();
-            this.mantenibilidadInterna = new ArrayList();
-            this.mantenibilidadExterna = new ArrayList();
+            this.listaMetricas = new ArrayList();
+            this.listaSeleccion = new ArrayList();
+            this.listaEvaluacion = new ArrayList();
+            this.indiceListas = 0;
         }
 
-        private void inicializarSeleccion()
+        private void inicializarEvaluacion(ArrayList metricas, int idProposito)
         {
-            this.MTSfuncionalidadInterna = new ArrayList();
-            this.MTSfuncionalidadExterna = new ArrayList();
-            this.MTSusabilidadInterna = new ArrayList();
-            this.MTSusabilidadExterna = new ArrayList();
-            this.MTSmantenibilidadInterna = new ArrayList();
-            this.MTSmantenibilidadExterna = new ArrayList();
-        }
+            JMetrica metricaJson;
+            MTEvaluacion metricaEval;
 
-        private void inicializarEvaluacion()
-        {
-            this.MTEfuncionalidadInterna = new ArrayList();
-            this.MTEfuncionalidadExterna = new ArrayList();
-            this.MTEusabilidadInterna = new ArrayList();
-            this.MTEusabilidadExterna = new ArrayList();
-            this.MTEmantenibilidadInterna = new ArrayList();
-            this.MTEmantenibilidadExterna = new ArrayList();
-        }
-
-        // Crear listas de metricas por caracteristicas
-
-        private void cargarListaFuncionabilidad(JFuncionabilidad funcionalidad, ArrayList metricas)
-        {
-            for (int i = 0; i < funcionalidad.Adecuacion.Length; i++)
-                metricas.Add(funcionalidad.Adecuacion[i]);
-
-            for (int i = 0; i < funcionalidad.Exactitud.Length; i++)
-                metricas.Add(funcionalidad.Exactitud[i]);
-
-            for (int i = 0; i < funcionalidad.Interoperabilidad.Length; i++)
-                metricas.Add(funcionalidad.Interoperabilidad[i]);
-
-            for (int i = 0; i < funcionalidad.SeguridadAcceso.Length; i++)
-                metricas.Add(funcionalidad.SeguridadAcceso[i]);
-
-            for (int i = 0; i < funcionalidad.CumplimientoFuncional.Length; i++)
-                metricas.Add(funcionalidad.CumplimientoFuncional[i]);
-        }
-
-        private void cargarListaUsabilidad(JUsabilidad usabilidad, ArrayList metricas)
-        {
-            for (int i = 0; i < usabilidad.Comprensibilidad.Length; i++)
-                metricas.Add(usabilidad.Comprensibilidad[i]);
-
-            for (int i = 0; i < usabilidad.Aprendizaje.Length; i++)
-                metricas.Add(usabilidad.Aprendizaje[i]);
-
-            for (int i = 0; i < usabilidad.Operabilidad.Length; i++)
-                metricas.Add(usabilidad.Operabilidad[i]);
-
-            for (int i = 0; i < usabilidad.Atractividad.Length; i++)
-                metricas.Add(usabilidad.Atractividad[i]);
-
-            for (int i = 0; i < usabilidad.CumplimientoUsabilidad.Length; i++)
-                metricas.Add(usabilidad.CumplimientoUsabilidad[i]);
-        }
-
-        private void cargarListaMantenibilidad(JMantenibilidad mantenibilidad, ArrayList metricas)
-        {
-            for (int i = 0; i < mantenibilidad.Analizabilidad.Length; i++)
-                metricas.Add(mantenibilidad.Analizabilidad[i]);
-
-            for (int i = 0; i < mantenibilidad.Modificabilidad.Length; i++)
-                metricas.Add(mantenibilidad.Modificabilidad[i]);
-
-            for (int i = 0; i < mantenibilidad.Estabilidad.Length; i++)
-                metricas.Add(mantenibilidad.Estabilidad[i]);
-
-            for (int i = 0; i < mantenibilidad.Testeabilidad.Length; i++)
-                metricas.Add(mantenibilidad.Testeabilidad[i]);
-
-            for (int i = 0; i < mantenibilidad.CumplimientoMantenibilidad.Length; i++)
-                metricas.Add(mantenibilidad.CumplimientoMantenibilidad[i]);
-        }
-
-        private void cargarJsonMetricas()
-        {
-            if (isFunIntAct)
+            for (int i = 0; i < metricas.Count; i++)
             {
-                funInt = JsonConvert.DeserializeObject<JFuncionabilidad>(File.ReadAllText("../../Archivos_configuracion/FuncionalidadInterna.json"));
-                cargarListaFuncionabilidad(funInt, funcionalidadInterna);
-            }
+                metricaEval = new MTEvaluacion();
 
-            if (isFunExtAct)
-            {
-                funExt = JsonConvert.DeserializeObject<JFuncionabilidad>(File.ReadAllText("../../Archivos_configuracion/FuncionalidadExterna.json"));
-                cargarListaFuncionabilidad(funExt, funcionalidadExterna);
-            }
+                metricaJson = (JMetrica)metricas[i];
+                metricaEval.Id = metricaJson.Id;
+                metricaEval.Formula = metricaJson.Formula[idProposito];
+                metricaEval.Parametros = metricaJson.Parametros;
 
-            if (isUsaIntAct)
-            {
-                usaInt = JsonConvert.DeserializeObject<JUsabilidad>(File.ReadAllText("../../Archivos_configuracion/UsabilidadInterna.json"));
-                cargarListaUsabilidad(usaInt, usabilidadInterna);
-            }
+                for (int j = 0; j < metricaJson.Parametros.Length; j++)
+                {
+                    metricaEval.Valores[j] = 0f;
+                }
+                
+                metricaEval.MejorValor = metricaJson.Mejor_valor;
+                metricaEval.PeorValor = metricaJson.Peor_valor;
 
-            if (isUsaExtAct)
-            {
-                usaExt = JsonConvert.DeserializeObject<JUsabilidad>(File.ReadAllText("../../Archivos_configuracion/UsabilidadExterna.json"));
-                cargarListaUsabilidad(usaExt, usabilidadExterna);
-            }
-
-            if (isManIntAct)
-            {
-                manInt = JsonConvert.DeserializeObject<JMantenibilidad>(File.ReadAllText("../../Archivos_configuracion/MantenibilidadInterna.json"));
-                cargarListaMantenibilidad(manInt, mantenibilidadInterna);
-            }
-
-            if (isManExtAct)
-            {
-                manExt = JsonConvert.DeserializeObject<JMantenibilidad>(File.ReadAllText("../../Archivos_configuracion/MantenibilidadExterna.json"));
-                cargarListaMantenibilidad(manExt, mantenibilidadExterna);
+                listaSeleccion.Add(metricaEval);
             }
         }
 
-        // Cambia las letras y los colores de las etiquetas de estado
-
-        private void cambiarEstado(ColorEstado estado, Label etiqueta)
+        private void limpiarSlider()
         {
-            var bc = new BrushConverter();
+            lblParam0.Content = "";
+            lblParam0.Visibility = Visibility.Hidden;
+            txtParam0.IsEnabled = false;
+            txtParam0.Visibility = Visibility.Hidden;
+            sldparam0.IsEnabled = false;
+            sldparam0.Visibility = Visibility.Hidden;
 
-            etiqueta.Background = (Brush)bc.ConvertFrom(estado.Color);
-            etiqueta.Content = estado.Etiqueta;
+            lblParam1.Content = "";
+            lblParam1.Visibility = Visibility.Hidden;
+            txtParam1.IsEnabled = false;
+            txtParam1.Visibility = Visibility.Hidden;
+            sldparam1.IsEnabled = false;
+            sldparam1.Visibility = Visibility.Hidden;
+            
+            lblParam2.Content = "";
+            lblParam2.Visibility = Visibility.Hidden;
+            txtParam2.IsEnabled = false;
+            txtParam2.Visibility = Visibility.Hidden;
+            sldparam2.IsEnabled = false;
+            sldparam2.Visibility = Visibility.Hidden;
         }
 
-        // Eventos botones menu flotante (flyout)
+        // Cargar una metrica
 
-        private void btnTileClick(object sender, System.Windows.RoutedEventArgs e)
+        private void cargarMetrica(JMetrica metrica, int idProposito) 
         {
-            Button clickedButton = (Button)e.Source;
+            System.Text.StringBuilder sb = new System.Text.StringBuilder();
 
-            //Comprobar el estado de la seleccion de metricas
+            lblID.Content = metrica.Id;
+            lblSubcaracteristica.Content = metrica.Subcaracteristica;
+            lblNombre.Content = metrica.Nombre;
+            labelProposito.Text = metrica.Proposito[idProposito]; 
+            labelMetodo.Text = metrica.Metodo;
+            label1_formula.Content = metrica.Formula[idProposito]; 
+            lblMejorValor.Content = metrica.Mejor_valor;
+            lblPeorValor.Content = metrica.Peor_valor;
 
-            switch (clickedButton.Name)
+            limpiarSlider();
+
+            if (metrica.Parametros.Length == 1 || metrica.Parametros.Length > 1)
             {
-                case "btnFuncInterna":
-                    paginaEvaluacion.cargarEvaluacionMetricas(this, "Funcionalidad", "Interna", funcionalidadInterna, MTSfuncionalidadInterna, MTEfuncionalidadInterna);
-                    MTSfuncionalidadInterna = paginaEvaluacion.evaluacionMetrica();
-                break;
-
-                case "btnUsabInterna":
-                    paginaEvaluacion.cargarEvaluacionMetricas(this, "Usabilidad", "Interna", usabilidadInterna, MTSusabilidadInterna, MTEusabilidadInterna);
-                    MTSusabilidadInterna = paginaEvaluacion.evaluacionMetrica();
-                break;
-
-                case "btnMantInterna":
-                    paginaEvaluacion.cargarEvaluacionMetricas(this, "Mantenibilidad", "Interna", mantenibilidadInterna, MTSmantenibilidadInterna, MTEmantenibilidadInterna);
-                    MTSmantenibilidadInterna = paginaEvaluacion.evaluacionMetrica();
-                break;
-
-                case "btnFuncExterna":
-                    paginaEvaluacion.cargarEvaluacionMetricas(this, "Funcionalidad", "Externa", funcionalidadExterna, MTSfuncionalidadExterna, MTEfuncionalidadExterna);
-                    MTSfuncionalidadExterna = paginaEvaluacion.evaluacionMetrica();
-                break;
-
-                case "btnUsabExterna":
-                    paginaEvaluacion.cargarEvaluacionMetricas(this, "Usabilidad", "Externa", usabilidadExterna, MTSusabilidadExterna, MTEusabilidadExterna);
-                    MTSusabilidadExterna = paginaEvaluacion.evaluacionMetrica();
-                break;
-
-                case "btnMantExterna":
-                    paginaEvaluacion.cargarEvaluacionMetricas(this, "Mantenibilidad", "Externa", mantenibilidadExterna, MTSmantenibilidadExterna, MTEmantenibilidadExterna);
-                    MTSmantenibilidadExterna = paginaEvaluacion.evaluacionMetrica();
-                break;
-
-                default:
-                    paginaEvaluacion.cargarEvaluacionMetricas(this, "Funcionalidad", "Interna", funcionalidadInterna, MTSfuncionalidadInterna, MTEfuncionalidadInterna);
-                    MTSfuncionalidadInterna = paginaEvaluacion.evaluacionMetrica();
-                break;
+                lblParam0.Content = metrica.Parametros[0];
+                lblParam0.Visibility = Visibility.Visible;
+                txtParam0.IsEnabled = true;
+                txtParam0.Visibility = Visibility.Visible;
+                sldparam0.IsEnabled = true;
+                sldparam0.Visibility = Visibility.Visible;
+            }
+                
+            if (metrica.Parametros.Length == 2 || metrica.Parametros.Length > 2)
+            {
+                lblParam1.Content = metrica.Parametros[1];
+                lblParam1.Visibility = Visibility.Visible;
+                txtParam1.IsEnabled = true;
+                txtParam1.Visibility = Visibility.Visible;
+                sldparam1.IsEnabled = true;
+                sldparam1.Visibility = Visibility.Visible;
+            }
+                
+            if (metrica.Parametros.Length == 3)
+            {
+                lblParam2.Content = metrica.Parametros[2];
+                lblParam2.Visibility = Visibility.Visible;
+                txtParam2.IsEnabled = true;
+                txtParam2.Visibility = Visibility.Visible;
+                sldparam2.IsEnabled = true;
+                sldparam2.Visibility = Visibility.Visible;
+            }
+                
+            for (int i = 0; i < metrica.Desc_param.Length; i++)
+            {
+                if (i != (metrica.Desc_param.Length - 1))
+                    sb.AppendLine(metrica.Desc_param[i] + "\n");
+                else
+                    sb.AppendLine(metrica.Desc_param[i]);
             }
 
-            this.NavigationService.Navigate(paginaEvaluacion);
+            txbkParam.Text = sb.ToString();
+        }
+
+        // Comprobar el estado de la evaluacion de la metrica
+
+        private void cargarEvaluacion(int indice)
+        {
+            MTEvaluacion metrica;
+            int parametros = 0;
+
+            metrica = (MTEvaluacion)listaEvaluacion[indice];
+            parametros = metrica.Parametros.Length;
+
+            if (parametros == 1)
+                txtParam0.Text = metrica.Valores[0].ToString();
+            
+            if (parametros == 2)          
+                txtParam1.Text = metrica.Valores[1].ToString();
+
+            if (parametros == 3)
+                txtParam2.Text = metrica.Valores[2].ToString();
+        }
+
+        // Guardar estado de la evaluacion de la metrica
+
+        private void guardarEvaluacion(int indice)
+        {
+            MTEvaluacion metrica;
+            int parametros = 0;
+
+            metrica = (MTEvaluacion)listaEvaluacion[indice];
+            parametros = metrica.Parametros.Length;
+
+            if (parametros == 1)            
+                metrica.Valores[0] = float.Parse(txtParam0.Text);           
+
+            if (parametros == 2)            
+                metrica.Valores[1] = float.Parse(txtParam1.Text);
+            
+            if (parametros == 3)           
+                metrica.Valores[2] = float.Parse(txtParam2.Text);
+            
+            listaEvaluacion[indice] = metrica;
+        }
+
+        // Retorna las metricas evaluadas
+
+        public ArrayList evaluacionMetrica()
+        {
+            return listaEvaluacion;
+        }
+
+        // Cargar la caracteristicas (Eventos botones tile)
+
+        public void cargarEvaluacionMetricas(FormularioEvaluacionPage llamada, string caracteristica, string perspectiva, ArrayList metricas, ArrayList seleccion, ArrayList evaluacion)
+        {
+            MTSeleccion metricaSeleccionada;
+
+            origen = llamada;
+            cargarEntorno();
+
+            // Etiquetas pricipales 
+            lblCaracteristica.Content = caracteristica;
+            lblPerspectiva.Content = perspectiva;
+
+            //Cargar listas metricas y seleccion
+            listaMetricas = (ArrayList)metricas.Clone();
+            listaSeleccion = (ArrayList)seleccion.Clone();
+            listaEvaluacion = (ArrayList)evaluacion.Clone();
+
+            metricaSeleccionada = (MTSeleccion)listaSeleccion[0];
+
+            //Si no hay seleccion previa, se crea
+            if (listaSeleccion.Count == 0)
+            {
+                inicializarEvaluacion(listaMetricas, metricaSeleccionada.Proposito);
+            }
+
+            //Cargo y compruebo metrica inicial
+            cargarMetrica((JMetrica)listaMetricas[0], metricaSeleccionada.Proposito);
+            cargarEvaluacion(0);
+        }
+
+        // Retroceder
+
+        private void retroceder(ref int indice, ArrayList metricas, ArrayList seleccionadas)
+        {
+            MTSeleccion metricaSeleccionada;
+
+            guardarEvaluacion(indice);
+
+            if ((indice - 1) > -1)
+            {
+                indice--;
+
+                if (indice == 0)
+                {
+                    btnAnterior.IsEnabled = false;
+                }
+
+                metricaSeleccionada = (MTSeleccion)seleccionadas[indice];
+
+                cargarMetrica((JMetrica)metricas[indice], metricaSeleccionada.Proposito);
+                cargarEvaluacion(indice);
+
+                if (btnSiguiente.IsEnabled == false)
+                {
+                    btnSiguiente.IsEnabled = true;
+                }
+            }
+            else
+            {
+                MessageBox.Show("No hay más metricas", "Aviso", System.Windows.MessageBoxButton.OK, System.Windows.MessageBoxImage.Exclamation);
+                btnAnterior.IsEnabled = false;
+            }
+        }
+
+        // Avanzar
+
+        private void avanzar(ref int indice, ArrayList metricas, ArrayList seleccionadas)
+        {
+            MTSeleccion metricaSeleccionada;
+
+            guardarEvaluacion(indice);
+
+            if ((indice + 1) < metricas.Count)
+            {
+                indice++;
+
+                if (indice == (metricas.Count - 1))
+                {
+                    btnSiguiente.IsEnabled = false;
+                }
+
+                metricaSeleccionada = (MTSeleccion)seleccionadas[indice];
+
+                cargarMetrica((JMetrica)metricas[indice], metricaSeleccionada.Proposito);
+                cargarEvaluacion(indice);
+
+                if (btnAnterior.IsEnabled == false)
+                {
+                    btnAnterior.IsEnabled = true;
+                }
+            }
+            else
+            {
+                MessageBox.Show("No hay más metricas", "Aviso", System.Windows.MessageBoxButton.OK, System.Windows.MessageBoxImage.Exclamation);
+                btnSiguiente.IsEnabled = false;
+            }
+        }
+
+        // Eventos Sliders
+
+        private void sldparam0_ValueChanged(object sender, RoutedPropertyChangedEventArgs<double> e)
+        {
+            txtParam0.Text = sldparam0.Value.ToString();
+        }
+
+        private void sldparam1_ValueChanged(object sender, RoutedPropertyChangedEventArgs<double> e)
+        {
+            txtParam1.Text = sldparam1.Value.ToString();
+        }
+
+        private void sldparam2_ValueChanged(object sender, RoutedPropertyChangedEventArgs<double> e)
+        {
+            txtParam2.Text = sldparam2.Value.ToString();
+        }
+
+        // Eventos botones
+
+        private void btnAnterior_Click(object sender, System.Windows.RoutedEventArgs e)
+        {
+            retroceder(ref indiceListas, listaMetricas, listaSeleccion);
+        }
+
+        private void btnSiguiente_Click(object sender, System.Windows.RoutedEventArgs e)
+        {
+            avanzar(ref indiceListas, listaMetricas, listaSeleccion);
+        }
+
+		private void btnGuardar_Click(object sender, RoutedEventArgs e)
+		{
+            guardarEvaluacion(indiceListas);
+            NavigationService.Navigate(origen);
+        }
+
+        private void btnTerminar_Click(object sender, RoutedEventArgs e)
+        {
+            guardarEvaluacion(indiceListas);
+            NavigationService.Navigate(origen);
         }
     }
 }
